+++ src/main/dll/CC/dll_0122_cctestinfot.c
@@ -35,7 +35,7 @@ void cctestinfot_init(int obj, s8 *def)
 {
     u32 v;
     v = (u32)((GameObject*)obj)->objectFlags | 0x6000;
-    ((GameObject*)obj)->objectFlags = (u16)v;
+    ((GameObject*)obj)->objectFlags = v;
     ((GameObject*)obj)->anim.rotX = (s16)((s32)(u8)def[0x1A] << 8);
     ((GameObject*)obj)->anim.rotY = (s16)((s32)(u8)def[0x19] << 8);
     ((GameObject*)obj)->anim.rotZ = (s16)((s32)(u8)def[0x18] << 8);
